Suggest catalogs, database and tables one-by-one, hierarchically
So instead of showing `catalog1.schema1.table1` all at once after FROM, we'll only show catalog1, and then after a dot, we'll suggest schema1, and only when we have `SELECT * FROM catalog1.schema1.` will be suggest table1.

It's a total hack, putting it all in the `createTableCandidates` function, but it works and it's the fastest way to implement this.